% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_section.R
\name{create_section}
\alias{create_section}
\title{Create a code section}
\usage{
create_section(sec_name)
}
\arguments{
\item{sec_name}{The name of the code section}
}
\value{
A character in the form of an RStudio code section
}
\description{
This function mimics the built in RStudio shortcut for creating sections.
It is used as part of the function set up the code documentation skeleton.
}
