% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openeR.R
\name{openeR}
\alias{openeR}
\title{Inserts an empty doc template}
\usage{
openeR()
}
\value{
A doc template
}
\description{
Inserts a doc template at the cursor position.
Should be used by mapping it to a shortcut in Rstudio. To do this,
go into 'Tools > Modify Keyboard Shortcuts' and then search for
'Insert doc skeleton'.
The template includes the project name, the content of the script,
one or more contributors, when the script was last updated and a
brief summary.
}
