% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openeR_gadget.R
\name{openeR_gadget}
\alias{openeR_gadget}
\title{Document your script with user specified doc skeletons}
\usage{
openeR_gadget()
}
\value{
A user specified doc skeleton
}
\description{
Opens a dialogue window with text inputs for the project ID,
the type content (e.g. data preparation), the contributors and a summary of
what the script does and how it can be used. Once the 'Done' button is clicked,
the skeleton is rendered as part of the script at the cursor position.
The function makes use of openeR::doc_skeleton() and the packages miniUI and shiny.
Should be used by mapping it to a shortcut in Rstudio. To do this,
go into 'Tools > Modify Keyboard Shortcuts' and then search for
'Insert a user specified doc skeleton'.
}
